Renamed Color to ColorSwatch
