+++ src/core/ColorPicker/ColorSwatch.tsx
